Only delete proxy-repo if exists

+++ testsuite/legacy-testsuite/src/test/java/org/sonatype/nexus/testsuite/p2/AbstractNexusProxyP2IT.java
@@ -89,7 +89,10 @@ public abstract class AbstractNexusProxyP2IT
 
         final File dest = new File( localStorageDir );
 
-        FileUtils.forceDelete( dest );
+        if ( dest.exists() )
+        {
+            FileUtils.forceDelete( dest );
+        }
 
         File source = getTestResourceAsFile( "proxy-repo" );
         if ( !source.exists() )
